Refactor config into system config within module
