+++ astoria/common/config/__init__.py
@@ -1,0 +1,5 @@
+"""Configuration schemas for Astoria."""
+
+from .system import AstoriaConfig
+
+__all__ = ["AstoriaConfig"]
