+++ astoria/common/mqtt/wrapper.py
@@ -18,7 +18,7 @@ from uuid import UUID
 import gmqtt
 from pydantic import BaseModel
 
-from astoria.common.config import MQTTBrokerInfo
+from astoria.common.config.system import MQTTBrokerInfo
 from astoria.common.ipc import ManagerMessage, ManagerRequest, RequestResponse
 
 from .topic import Topic
